Cleanup from implicit usings
