+++ src/EmployeeETL.WebAPI/EmployeeETL.WebAPI.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <TargetFramework>net7.0</TargetFramework>
     <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
+    <!-- <ImplicitUsings>enable</ImplicitUsings> -->
   </PropertyGroup>
 
   <ItemGroup>
