+++ src/EmployeeETL.WebAPI/OpenApi/OpenApiConfiguration.cs
@@ -1,7 +1,8 @@
-using System.Runtime.CompilerServices;
-using Asp.Versioning.ApiExplorer;
+using System.Linq;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Options;
-using Microsoft.OpenApi.Models;
 using Swashbuckle.AspNetCore.SwaggerGen;
 
 namespace EmployeeETL.WebAPI.OpenApi;
@@ -34,28 +35,5 @@ public static class OpenApiConfigurations
         }
 
         return app;
-    }
-}
-
-internal class SwaggerGenConfigurationOptions : IConfigureOptions<SwaggerGenOptions>
-{
-    private readonly IApiVersionDescriptionProvider _provider;
-
-    public SwaggerGenConfigurationOptions(IApiVersionDescriptionProvider provider)
-    {
-        _provider = provider;
-    }
-
-    public void Configure(SwaggerGenOptions options)
-    {
-        _provider.ApiVersionDescriptions.ToList().ForEach(apiVersion =>
-        {
-            options.SwaggerDoc(apiVersion.GroupName, new OpenApiInfo()
-            {
-                Title = $"Employee ETL API {apiVersion.ApiVersion}",
-                Version = apiVersion.ApiVersion.ToString(),
-                Description = "Provides basic load features"
-            });
-        });
     }
 }
