+++ src/EmployeeETL.WebAPI/Versioning/VersionConfiguration.cs
@@ -1,4 +1,6 @@
 using Asp.Versioning;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.Extensions.DependencyInjection;
 
 namespace EmployeeETL.WebAPI.Versioning;
